add routes

+++ street.php
@@ -62,7 +62,7 @@
                                     <td><?= $data['address'] ?></td>
                                     <td>
                                         <a href="edit-street.php?id=<?php echo $data['id'] ?>" class="btn btn-warning ">edit</a>
-                                        <a href="delete-street.php?id=<?php echo $data['id'] ?>" class="btn btn-danger ">delete</a>
+                                        <a href="delete-street.php?id=<?php echo $data['id'] ?>" onclick="return confirm('Are You Sure ?')" class="btn btn-danger ">delete</a>
                                     </td>
                                 </tr>
                             <?php } ?>
